Collision Detection
Added Collision Detection. Collision Boxes can be different bases on the
mob. Fixed users being able to go into the "negative" regions of a map.
"Negative" regions are those outside the bounds of the map and such are
solid and are instances of the voidTile.

+++ src/com/burksnet/code/games/rain/console/ConsoleManager.java
@@ -73,7 +73,7 @@ public class ConsoleManager implements Runnable {
 				if (array.length > 2) {
 					throw new ArrayIndexOutOfBoundsException(array.length - 1);
 				}
-				double local1 = Double.parseDouble(array[1]);
+				int local1 = Integer.parseInt(array[1]);
 				System.out.println("User Speed changed.");
 				game.player.speed = local1;
 			} catch (NumberFormatException e) {
